multimessage: list of cores now given as argument
The idea here is to make the benchmark more flexible. For example, the
choice of cores to benchmark could be determined by the Simulator
based on the pairwise measurements.

To reduce undeterminism in the benchmark, sending to remote cores on
several NUMA nodes should be avoided. Cores on two different nodes
have potentially (and likely) completely different performance
characteristics, which can make the result of the multimessage hard to
read.

This is an attempt to make more sense of the heat maps generated by
the multi message benchmark.