Revert "nsend: added warm-up phase"
The TSC difference was wrongly calculated with that patch. The newly
added warmup phase was included in the measurement. With the evaluation
script still dividing the result by the BATCH_SIZE, this caused the send
time to be 2x off.

While adding a warmup phase might be something worth looking at, I am
reverting this patch for now. Pairwise has been executed with the
version before this patch on pretty much all machines, so my first
priority is to keep them consistent.

This reverts commit ce45deb0eded659205716bce0c4aa4e879e12615.